Styling

+++ app/models/payment.rb
@@ -6,5 +6,5 @@ class Payment < ApplicationRecord
   validates :name, presence: true
   validates :amount, presence: true, numericality: { greater_than_or_equal_to: 0 }
 
-  attribute :amount, :integer, default: 0
+  attribute :amount, :decimal, default: 0
 end
